Add first repositories and er diagram

+++ domain/src/main/java/artist/Artist.java
@@ -4,19 +4,19 @@ import java.util.Date;
 
 public class Artist {
 
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private Date dateOfBirth;
 
-    public Artist(int id, String firstName, String lastName, Date dateOfBirth) {
+    public Artist(long id, String firstName, String lastName, Date dateOfBirth) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
         this.dateOfBirth = dateOfBirth;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
